Fix SimpleImageComparator for different sized images
If SimpleImageComparator is provided with two differently
sized images, it throws an exception due to calling getPixel() on
with out-of-bounds coordinates. This PR fixes that by adding some
bounds checking to various places where getPixel() is called.

I also added some checking of the diff mask size, as it's currently
possible to provide an incorrect sized mask, which would then
cause an out-of-bounds exception too.